refactor: rename DiffModal to CompareModal
