+++ src/components/organisms/CompareModal/CompareModal.styled.tsx
