+++ src/components/organisms/CompareModal/CompareState.ts
