+++ src/components/organisms/CompareModal/__test__/fixtures/basicDeployment.ts
