--- src/components/organisms/DiffModal/index.ts
@@ -1,1 +1,0 @@
-export {default} from './DiffModal';
